nodes/interrupts: Moved interrupts.c from board_driver to nodes
When compiled into the board_driver lib some ISRs are optimized out
on some systems. This is avoided by moving it to root/nodes/
and adding it to our custom target setup in cmake,
in the same way (and for a simular reason) as newlib_calls.c
With this setup interrupts.c is automatically compiled into all targets.