board_driver/usb/usbd_cdc_interface: Cleaned interface of uart related code.
The old interface taken from example used UART to echo data to and from
the USB interface. This is not what we want.
Also it was missing a piece of code for correctly receiving data,
which is added in this commit usbd_cdc_interface.c function CDC_Itf_Receive
line 56. This line tells the driver that we are ready to recieve another
packet. Insert additional code above to save received data.