nodes/newlib nodes/examples/usb: Working USB for serial communication
With this commit sending/receiving serial data across the cdc USB interface
is working and can also be used with stdio calls like printf (although this
is deactivated by default). In contrast to the UART stdio interface, the USB
interface uses solely \n instead of \r\n and any terminal emulator should be
setup as such.